+++ ChangeLog.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Irrep.generators` and `Irreps.generators` functions to get the generators of the representations.
 - `e3nn.bessel` function
+- `slice_by_mul`, `slice_by_dim` and `slice_by_chunk` functions to `Irreps` and `IrrepsArray`
 
 ### Changed
 - **breaking change** `e3nn.soft_one_hot_linspace` does not support `bessel` anymore. Use `e3nn.bessel` instead.
+
+### Removed
+- **breaking change** `IrrepsArray.split`
 
 ## [0.9.2] - 2022-09-29
 ### Fixed